Use logRecord instead of log, to prevent stringifying the message
Logger.log creates a LogRecord under the hood, however it calls String() on its
argument, so we lose the richness of ClojureScript datatypes, which in turn
means tools like cljs-devtools become a lot less useful.

To prevent Logger.log from doing this we forced it to take an alternative branch
where it does not call String(), by enabling a LogBuffer. However this approach
did not work with advanced compilation, and is hackish at best.

Instead we can drop to a lower level API, create the LogRecord instance
ourselves, and bypass the String() call in Logger.log.