Fix bug and add workaround in glogc.
Accept Arne's suggestion to put expressions in a list.

It's not ideal, but passing values to spy in a cljc file will work, even
if the way it appears is harder to read.

Also fixes a bug where the macro would generate `glogi/spy` for Clojure,
rather than `pedestal/spy`.

Co-authored-by: Arne Brasseur <arne@arnebrasseur.net>